Another change needed, to complete the new type of domDoc->documentNumber.

+++ generic/dom.h
@@ -125,7 +125,7 @@
 # define TDomThreaded(x)    x
 # define HASHTAB(doc,tab)   (doc)->tab
 # define NODE_NO(doc)       ((doc)->nodeCounter)++
-# define DOC_NO(doc)        (unsigned int)(doc)
+# define DOC_NO(doc)        (unsigned long)(doc)
 #endif /* TCL_THREADS */
 
 #define DOC_CMD(s,doc)      sprintf((s), "domDoc%p", (doc))
